requirements.txt
altair==5.5.0
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
executing==2.2.0
filelock==3.13.1
fonttools==4.56.0
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
idna==3.10
imageio==2.37.0
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.10.1
matplotlib-inline==0.1.7
mpmath==1.3.0
narwhals==1.32.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.4
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.7
plotly==6.0.1
prompt_toolkit==3.0.50
protobuf==5.29.4
psutil==7.0.0
pure_eval==0.2.3
pyarrow==19.0.1
pydeck==0.9.1
pydicom==3.0.1
Pygments==2.19.1
pyodbc==5.2.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pywin32==310
pyzmq==26.3.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
scikit-image==0.25.2
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.40
stack-data==0.6.3
streamlit==1.44.0
streamlit-browser-session-storage==0.0.12
streamlit-drawable-canvas @ git+https://github.com/SunOner/streamlit-drawable-canvas.git@01becf65b50c444cd524366b9230fd49ee2daebf
streamlit-image-zoom==0.0.4
streamlit-javascript==0.1.5
streamlit-js-eval==0.1.7
streamlit-local-storage==0.0.25
streamlit-screen-stats==0.0.82
sympy==1.13.1
tenacity==9.0.0
tifffile==2025.3.30
toml==0.10.2
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchvision==0.21.0+cu124
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
watchdog==6.0.0
wcwidth==0.2.13
